pyproject.toml
[project]
name = "csrf-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "daphne>=4.2.1",
    "django>=6.0.1",
    "django-browser-reload>=1.21.0",
    "django-eventstream>=5.3.3",
    "django-redis>=6.0.0",
    "pillow>=12.1.0",
]
